[SP-2051] - Backport of BISERVER-12637 - Jackrabbit's ds_repos_datastore table can grow out of control causing performance issues (5.4 Suite) - move the listener to the extensions module - revert changes of PentahoAuditLog.log, that were committed accidentally - revert changes in scheduler's ivy.xml
