+++ scheduler/ivy.xml
@@ -22,12 +22,10 @@
     <dependency org="junit" name="junit" rev="4.4" conf="test->default" />
     <dependency org="org.jmock" name="jmock-junit4" rev="2.5.1" conf="test->default" />
     	<dependency org="org.jmock" name="jmock-legacy" rev="2.5.1" conf="test->default" />
-    <dependency org="org.mockito" name="mockito-all"  rev="1.8.5" conf="test->default"/>
 
     <!--  internal dependencies -->
     <dependency org="${ivy.artifact.group}" name="pentaho-platform-api"  rev="${project.revision}" changing="true" conf="default->default" />
     <dependency org="${ivy.artifact.group}" name="pentaho-platform-core" rev="${project.revision}" changing="true" conf="default->default" />
-    <dependency org="${ivy.artifact.group}" name="pentaho-platform-repository" rev="${project.revision}" changing="true" conf="default->default" />
     <dependency org="pentaho" name="pentaho-versionchecker" rev="${dependency.pentaho-versionchecker.revision}" changing="true" />
    	<dependency org="javax.mail" name="mail" rev="1.4.1" transitive="false" />
